social_change_backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
anthropic==0.7.8
SQLAlchemy==2.0.23
Werkzeug==2.3.7
requests==2.31.0
gunicorn==21.2.0
python-dateutil==2.8.2 